Cleaned up main and made separate headers
Still TODO: Making a lobby screen for both host game and join game, replacing Debug text with AddUIElement, clean up usage of strings
