+++ TeamProject/MainMenuScreen.h
@@ -1,61 +1,63 @@
+#pragma once
+
 //#include "PushdownMachine.h"
 #include "PushdownState.h"
 #include <TutorialGame.h>
-#include "Window.h"
+//#include "Window.h"
 #include <CSC8503CoreClasses/Debug.h>
 #include <iostream>
+#include <GameScreen.h>
 
 using namespace NCL;
 using namespace NCL::CSC8503;
 
 class MainMenuScreen : public PushdownState {
+    int selection = 0;
+    bool inMenu;
+    TutorialGame* game;
+    Controller* controller;
+    std::string gameMode;
+
 public:
-	MainMenuScreen(TutorialGame* game) : game(game), selection(0), inMenu(true) {}
+    MainMenuScreen(Controller* controller, TutorialGame* game) : controller(controller), game(game), selection(0), inMenu(true) {}
 
-	PushdownResult OnUpdate(float dt, PushdownState** newState) override {
-		//if (!inMenu) return PushdownResult::Push; // Move to GameScreen when a mode is selected
+    PushdownResult OnUpdate(float dt, PushdownState** newState) override {
 
-		const std::string menuItems[3] = { "Singleplayer", "Host Game", "Join Game" };
+        const std::string menuItems[3] = { "Singleplayer", "Host Game", "Join Game" };
 
-		// Handle input
-		if (Window::GetKeyboard()->KeyPressed(NCL::KeyCodes::DOWN)) {
-			selection = std::min(2, selection + 1);
-		}
-		if (Window::GetKeyboard()->KeyPressed(NCL::KeyCodes::UP)) {
-			selection = std::max(0, selection - 1);
-		}
-		if (Window::GetKeyboard()->KeyPressed(NCL::KeyCodes::RETURN)) {
-			GameMode mode = static_cast<GameMode>(selection);
+        if (controller->GetDigital(Controller::DigitalControl::MenuDown)) {
+            selection = std::min(2, selection + 1);
+        }
+        if (controller->GetDigital(Controller::DigitalControl::MenuUp)) {
+            selection = std::max(0, selection - 1);
+        }
+        if (controller->GetDigital(Controller::DigitalControl::MenuConfirm)) {
+            GameMode mode = static_cast<GameMode>(selection);
 
-			if (mode == GameMode::SINGLEPLAYER) {
-				game->LoadWorldFromFile(8);
-			}
-			else {
-				game->JoinGame(mode == GameMode::HOST_GAME);
-			}
+            if (mode == GameMode::SINGLEPLAYER) {
+                game->LoadWorldFromFile(9);
+                gameMode = "Singleplayer";
+            }
+            else {
+                game->JoinGame(mode == GameMode::HOST_GAME);
+                gameMode = "Multiplayer";
+            }
 
-			inMenu = false;  // Hide menu once a game mode is selected
-			*newState = new GameScreen();  // Push the game state
-			return PushdownResult::Push;
-		}
+            *newState = new GameScreen(controller, game, gameMode);
+            inMenu = false;
+            return PushdownResult::Push;
+        }
 
-		// Render menu
-		for (int i = 0; i < 3; i++) {
-			std::string currentItem = menuItems[i];
-			if (i == selection) currentItem += " <";
-			Debug::Print(currentItem, Vector2(1, 50 + (10 * i)));
-		}
+        //Render menu
+        for (int i = 0; i < 3; i++) {
+            std::string currentItem = menuItems[i];
+            if (i == selection) currentItem += " <";
+            Debug::Print(currentItem, Vector2(1, 50 + (10 * i))); //Replace with AddUIElement
+        }
+        return PushdownResult::NoChange;
+    }
 
-		return PushdownResult::NoChange;
-	}
+    void OnAwake() override {
 
-	void OnAwake() override {
-		std::cout << "Main Menu Active\n";
-		inMenu = true;
-	}
-	
-protected:
-	TutorialGame* game;
-	int selection;
-	bool inMenu;
+    }
 };
