--- TeamProject/MainMenuState.h
@@ -1,1 +1,0 @@
-#pragma once
